removed pass, moved memmap to own file
